Add time_message_printed
This also combines some refactoring using mapped sqlalchemy instead of columns.
I've made the change because it's the more modern way of doing it and plays better
together with mypy. It also allows for optional fields, which is the main reason it's
part of this commit and not a seperate refactor commit, as the time_message_printed
can be null (aka: not printed yet!)